pyproject.toml
[tool.poetry]
name = "jidou_hikki"
version = "0.1.0"
description = ""
authors = ["fariz.tumbuan <fariz.tumbuan@gmail.com>"]
license = "MIT"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
SudachiPy = "^0.5.2"
django-model-utils = "^4.1.1"
jaconv = "^0.3"
SudachiDict-full = {version = "^20210608", extras = ["production"]}
jamdict = "^0.1a11"
jamdict-data = "^1.5"
gunicorn = "^20.1.0"
pydantic = "^1.9.1"
django-ninja = "^0.19.1"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.dephell.main]    
from = {format = "poetry", path = "pyproject.toml"}    
to = {format = "setuppy", path = "setup.py"}
